Rename and build on project-plugin
