+++ .travis.yml
@@ -1,31 +1,1 @@
-language: java
-jdk:
-- openjdk11
-
-env:
-- TERM=dumb
-
-#Below skips the installation step completely (https://docs.travis-ci.com/user/customizing-the-build/#Skipping-the-Installation-Step)
-#We need it because otherwise Travis CI injects an awkward './gradlew assemble' step into the CI workflow
-#We want to control and decide what Gradle tasks are executed
-install:
-- true
-
-before_script:
-- git config --global user.email "SvcReleaserBackend@yoomoney.ru"
-- git config --global user.name "YoomoneyRobot"
-
-script:
-- ./gradlew -b build.gradle clean build bintrayUpload publishPlugins --info  -D"gradle.publish.key"=$GRADLE_PUBLISH_KEY -D"gradle.publish.secret"=$GRADLE_PUBLISH_SECRET
-
-after_success:
-- bash <(curl -s https://codecov.io/bash)
-- ./publish_doc.sh
-
-before_cache:
-- rm -f  $HOME/.gradle/caches/modules-2/modules-2.lock
-- rm -fr $HOME/.gradle/caches/*/plugin-resolution/
-cache:
-  directories:
-  - $HOME/.gradle/caches/
-  - $HOME/.gradle/wrapper/
+import: yoomoney-gradle-plugins/travis-shared-configuration:build-and-publish-plugin.yml@master
