+++ coverage.properties
@@ -1,0 +1,4 @@
+instruction=87
+branch=62
+method=95
+class=100
