+++ project.gradle
@@ -1,0 +1,8 @@
+repositories {
+    mavenCentral()
+    maven { url 'https://plugins.gradle.org/m2/' }
+
+    dependencies {
+        classpath 'ru.yoomoney.gradle.plugins:gradle-project-plugin:7.+'
+    }
+}
