+++ src/main/java/ru/yoomoney/gradle/plugins/grafana/CollectGrafanaDashboardsTask.java
@@ -1,4 +1,4 @@
-package ru.yoomoney.tech.plugins.grafana;
+package ru.yoomoney.gradle.plugins.grafana;
 
 import org.gradle.api.DefaultTask;
 import org.gradle.api.artifacts.Configuration;
@@ -6,10 +6,10 @@ import org.gradle.api.logging.Logger;
 import org.gradle.api.logging.Logging;
 import org.gradle.api.tasks.TaskAction;
 import org.json.JSONObject;
-import ru.yoomoney.tech.plugins.grafana.impl.GrafanaDashboard;
-import ru.yoomoney.tech.plugins.grafana.impl.GrafanaDashboardCollector;
-import ru.yoomoney.tech.plugins.grafana.impl.KotlinScriptContentCreator;
-import ru.yoomoney.tech.plugins.grafana.impl.RawContentCreator;
+import ru.yoomoney.gradle.plugins.grafana.impl.GrafanaDashboard;
+import ru.yoomoney.gradle.plugins.grafana.impl.GrafanaDashboardCollector;
+import ru.yoomoney.gradle.plugins.grafana.impl.KotlinScriptContentCreator;
+import ru.yoomoney.gradle.plugins.grafana.impl.RawContentCreator;
 
 import java.io.File;
 import java.nio.file.Paths;
