+++ src/main/java/ru/yoomoney/gradle/plugins/grafana/impl/DashboardContentCreator.java
@@ -1,4 +1,4 @@
-package ru.yoomoney.tech.plugins.grafana.impl;
+package ru.yoomoney.gradle.plugins.grafana.impl;
 
 import javax.annotation.Nonnull;
 import java.io.File;
