+++ static-analysis.properties
@@ -1,0 +1,4 @@
+findbugs=9
+compiler=0
+checkstyle=0
+architecture=0
